--- arcusd/models/enums/opstatus.py
@@ -1,7 +1,0 @@
-from enum import Enum
-
-
-class OperationStatus(Enum):
-    NONE = 0
-    SUCCESS = 1
-    FAILED = 2
