--- arcusd/models/enums/optype.py
@@ -1,7 +1,0 @@
-from enum import Enum
-
-
-class OperationType(Enum):
-    QUERY = 1
-    PAYMENT = 2
-    TOPUP = 3
