+++ arcusd/models/operationinfo.py
@@ -1,11 +1,10 @@
-from .enums.optype import OperationType
-from .enums.opstatus import OperationStatus
+from ..types import OperationStatus, OperationType
 
 
 class OpInfo:
     def __init__(self,
                  tran_type: OperationType,
-                 status: OperationStatus = OperationStatus.NONE,
+                 status: OperationStatus = OperationStatus.none,
                  operation=None,
                  error_message=None):
         self.type = tran_type
