+++ requirements.txt
@@ -1,2 +1,2 @@
-arcus
+arcus>=0.1.2,<0.2
 celery>=4.2,<4.3
