+++ tests/test_tasks.py
@@ -1,7 +1,6 @@
 import pytest
 
-from arcusd.models.enums.opstatus import OperationStatus
-from arcusd.models.enums.optype import OperationType
+from arcusd.types import OperationStatus, OperationType
 from arcusd.daemon.tasks import (
     topup, query_bill, pay_bill, cancel_transaction)
 
@@ -9,8 +8,8 @@ from arcusd.daemon.tasks import (
 @pytest.mark.vcr(cassette_library_dir='tests/cassettes/test_tasks')
 def test_query_bill():
     op_info = query_bill(40, '501000000007')
-    assert op_info.status == OperationStatus.SUCCESS
-    assert op_info.type == OperationType.QUERY
+    assert op_info.status == OperationStatus.success
+    assert op_info.type == OperationType.query
     assert op_info.operation.account_number == '501000000007'
     assert type(op_info.operation.balance) is float
 
@@ -25,8 +24,8 @@ def test_query_bill():
 ])
 def test_query_bill_failed(biller_id, account_number, expected_message):
     op_info = query_bill(biller_id, account_number)
-    assert op_info.type == OperationType.QUERY
-    assert op_info.status == OperationStatus.FAILED
+    assert op_info.type == OperationType.query
+    assert op_info.status == OperationStatus.failed
     assert (op_info.error_message == expected_message
             or op_info.error_message.startswith(expected_message))
 
@@ -34,8 +33,8 @@ def test_query_bill_failed(biller_id, account_number, expected_message):
 @pytest.mark.vcr(cassette_library_dir='tests/cassettes/test_tasks')
 def test_successful_payment():
     op_info = pay_bill(40, '501000000007')
-    assert op_info.type == OperationType.PAYMENT
-    assert op_info.status == OperationStatus.SUCCESS
+    assert op_info.type == OperationType.payment
+    assert op_info.status == OperationStatus.success
     assert type(op_info.operation.id) is int
     assert op_info.operation.status == 'fulfilled'
 
@@ -43,15 +42,15 @@ def test_successful_payment():
 @pytest.mark.vcr(cassette_library_dir='tests/cassettes/test_tasks')
 def test_failed_payment():
     op_info = pay_bill(37, '2424240024')
-    assert op_info.type == OperationType.PAYMENT
-    assert op_info.status == OperationStatus.FAILED
+    assert op_info.type == OperationType.payment
+    assert op_info.status == OperationStatus.failed
 
 
 @pytest.mark.vcr(cassette_library_dir='tests/cassettes/test_tasks')
 def test_successful_topup():
     op_info = topup(13599, '5599999999', 100.0, 'MXN')
-    assert op_info.type == OperationType.TOPUP
-    assert op_info.status == OperationStatus.SUCCESS
+    assert op_info.type == OperationType.topup
+    assert op_info.status == OperationStatus.success
 
 
 @pytest.mark.vcr(cassette_library_dir='tests/cassettes/test_tasks')
@@ -61,8 +60,8 @@ def test_successful_topup():
 ])
 def test_failed_topup(phone_number, amount, expected_message):
     op_info = topup(13599, phone_number, amount, 'MXN')
-    assert op_info.type == OperationType.TOPUP
-    assert op_info.status == OperationStatus.FAILED
+    assert op_info.type == OperationType.topup
+    assert op_info.status == OperationStatus.failed
     assert op_info.error_message == expected_message
 
 
@@ -70,6 +69,6 @@ def test_failed_topup(phone_number, amount, expected_message):
 def test_cancel_bill():
     pay_op_info = pay_bill(35, '123456851236')
     cancel_op_info = cancel_transaction(pay_op_info.operation.id)
-    assert cancel_op_info.status == OperationStatus.SUCCESS
+    assert cancel_op_info.status == OperationStatus.success
     assert cancel_op_info.operation.id == pay_op_info.operation.id
     assert cancel_op_info.operation.status == 'refunded'
